Added files via upload
In Sri Lanka, the native language of the hearing impaired community is Sri Lankan sign language, which defines a set of vocabulary using gestures corresponding to frequently used words. If a word is not defined, it is spelt out using gestures that correspond to the letters in the Sinhala, Tamil or English alphabet.  

The Leap Motion is an innovative, 3D motion capturing device designed especially for hands and fingers tracking with precision up to 0.01mm. The outcome of the project is the Gesture Recognition System Using Leap Motion dedicated to the Sri Lankan English Sign Language. The authors examined the data provided by the sensor in context of recognition of hand poses (static gestures), hand movements (dynamic gestures) and in task of a finger recognition. The static gestures are recognized using the Support Vector Machine (SVM), with median filtering an input data and using the correspondences between consecutive recognitions. The dynamic gestures are recognized using the Hid